requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
chacha20poly1305-reuseable==0.12.0
charset-normalizer==3.3.2
cryptography==41.0.7
frozenlist==1.4.1
idna==3.6
ifaddr==0.2.0
mediafile==0.12.0
miniaudio==1.59
multidict==6.0.4
mutagen==1.47.0
pillow==10.2.0
protobuf==4.25.2
pyatv==0.14.5
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
requests==2.31.0
six==1.16.0
srptools==1.0.1
tabulate==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
yarl==1.9.4
zeroconf==0.131.0
